+++ python/gherkin/stream/source_events.py
@@ -1,10 +1,8 @@
-import io
-
 def source_event(path):
     event = {
         'source': {
             'uri': path,
-            'data': io.open(path, 'r', encoding='utf8', newline='').read(),
+            'data': open(path, encoding='utf8', newline='').read(),
             'mediaType': 'text/x.cucumber.gherkin+plain'
         }
     }