+++ CHANGELOG.md
@@ -19,6 +19,9 @@ This document is formatted according to the principles of [Keep A CHANGELOG](htt
 - [.NET] Enable warnings as errors
 - [Ruby] Initial rubocop autofixes (Mainly to style) ([#266](https://github.com/cucumber/gherkin/pull/266))
 
+### Removed
+- [Python] Drop compatibility for python 2. Supported python versions are 3.8, 3.9, 3.10, 3.12
+
 ## [29.0.0] - 2024-08-12
 ### Added
 - (i18n) Added Gujarati translation for "Rule" ([#249](https://github.com/cucumber/gherkin/pull/249))